Fix landblock loading at edges of map.

+++ include/LandcellId.h
@@ -35,14 +35,14 @@ class LandcellId
 public:
     LandcellId();
     explicit LandcellId(uint32_t value);
-    LandcellId(uint8_t x, uint8_t y);
-    LandcellId(uint8_t x, uint8_t y, uint16_t n);
+    LandcellId(int x, int y);
+    LandcellId(int x, int y, int n);
 
     int calcSquareDistance(LandcellId other) const;
 
-    uint8_t x() const;
-    uint8_t y() const;
-    uint16_t n() const;
+    int x() const;
+    int y() const;
+    int n() const;
     uint32_t value() const;
     bool isStructure() const;
